JUNITBENCH-33: Fixed extra embedded classes problem. Only _Double is now embedded.
Universal backport package name is replaced with an explicit method name now and it seems
to do the job. A runtime test would have to be executed to make sure no other JDK1.6 methods
are actually used (this Double.valueOf is implicitly inserted by the compiler, so more
trickery may lurk out there).